+++ src/onegov.town/HISTORY.rst
@@ -7,6 +7,9 @@ Unreleased
 - Adds e-mail notifications for open/close ticket.
   [href]
 
+- Adds reopen ticket functionality.
+  [msom]
+
 0.2.4 (2015-07-14)
 ~~~~~~~~~~~~~~~~~~~